backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
axios
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
scipy>=1.10.0
shap>=0.42.0
fairlearn>=0.9.0
groq
langchain
langchain-groq
langgraph
python-multipart
aiofiles
joblib
tqdm
